Split builder into several files
