+++ nengo_loihi/builder/__init__.py
@@ -1,0 +1,7 @@
+from .builder import (
+    Builder,
+    INTER_N,
+    INTER_RATE,
+    Model,
+)
+from . import connection, ensemble, node, probe
